requirements.txt
Flask==2.3.3
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.8.1.78
PyMuPDF==1.23.8
Werkzeug==2.3.7 